requirements.txt
pyyaml
pytest
beautifultable

